chore: dedupe evm_spec fn
