+++ cli/src/cmd/cast/run.rs
@@ -17,6 +17,7 @@ use forge::{
     utils::h256_to_b256,
 };
 use foundry_config::{find_project_root_path, Config};
+use foundry_evm::utils::evm_spec;
 use std::{collections::BTreeMap, str::FromStr};
 use tracing::trace;
 use ui::{TUIExitReason, Tui, Ui};
@@ -89,9 +90,8 @@ impl RunArgs {
 
         // configures a bare version of the evm executor: no cheatcode inspector is enabled,
         // tracing will be enabled only for the targeted transaction
-        let builder = ExecutorBuilder::default()
-            .with_config(env)
-            .with_spec(crate::utils::evm_spec(&config.evm_version));
+        let builder =
+            ExecutorBuilder::default().with_config(env).with_spec(evm_spec(&config.evm_version));
 
         let mut executor = builder.build(db);
