+++ cli/src/cmd/forge/script/executor.rs
@@ -1,15 +1,12 @@
 use super::*;
-use crate::{
-    cmd::{
-        ensure_clean_constructor,
-        forge::script::{
-            artifacts::ArtifactInfo,
-            runner::SimulationStage,
-            transaction::{AdditionalContract, TransactionWithMetadata},
-        },
-        needs_setup,
+use crate::cmd::{
+    ensure_clean_constructor,
+    forge::script::{
+        artifacts::ArtifactInfo,
+        runner::SimulationStage,
+        transaction::{AdditionalContract, TransactionWithMetadata},
     },
-    utils,
+    needs_setup,
 };
 use cast::executor::inspector::cheatcodes::util::BroadcastableTransactions;
 use ethers::{
@@ -23,6 +20,7 @@ use forge::{
     CallKind,
 };
 use foundry_common::{shell, RpcUrl};
+use foundry_evm::utils::evm_spec;
 use futures::future::join_all;
 use parking_lot::RwLock;
 use std::{collections::VecDeque, sync::Arc};
@@ -306,7 +304,7 @@ impl ScriptArgs {
 
         let mut builder = ExecutorBuilder::default()
             .with_config(env)
-            .with_spec(utils::evm_spec(&script_config.config.evm_version))
+            .with_spec(evm_spec(&script_config.config.evm_version))
             .with_gas_limit(script_config.evm_opts.gas_limit())
             // We need it enabled to decode contract names: local or external.
             .set_tracing(true);
